A lot of changes

+++ src/main/java/dev/u9g/minigames/games/gathering/util/GatheringWorld.kt
@@ -2,12 +2,14 @@ package dev.u9g.minigames.games.gathering.util
 
 import dev.u9g.minigames.util.Task
 import dev.u9g.minigames.util.mm
+import dev.u9g.minigames.util.throwablerenderer.sendToOps
+import net.kyori.adventure.util.Ticks
 import org.bukkit.Bukkit
-import org.bukkit.OfflinePlayer
 import org.bukkit.World
 import org.bukkit.WorldCreator
 import org.bukkit.entity.Player
-import java.util.UUID
+import org.bukkit.event.player.PlayerTeleportEvent
+import java.util.*
 import java.util.concurrent.CompletableFuture
 import java.util.function.Consumer
 
@@ -27,7 +29,12 @@ class GatheringWorld {
         val nonNullWorld = (world ?: throw Error("called GatheringWorld#spawnPoint with a null world"))
         val loc = nonNullWorld.getHighestBlockAt(0, 0).location
         nonNullWorld.getChunkAtAsync(loc, Consumer {
-            player.teleportAsync(loc).thenAccept {
+            player.sliver_teleportAsync(loc, PlayerTeleportEvent.TeleportCause.PLUGIN).thenAccept {
+                if (!it.isSuccessful) {
+                    val err = Exception("Player failed to teleport due to ${it.name}")
+                    err.sendToOps()
+                    err.printStackTrace()
+                }
                 cf.complete(null)
             }
         })
@@ -36,12 +43,16 @@ class GatheringWorld {
 
     fun delete() {
         val theWorld = world ?: throw Error("called GatheringWorld#delete with a null world")
-        Bukkit.unloadWorld(theWorld, false)
-        println("starting delete")
-        Task.asyncDelayed(0) {
-            if (!theWorld.worldFolder.deleteRecursively())
-                Bukkit.broadcast("failed to delete".mm())
-            println("done deleting")
+        val unloadResult = Bukkit.sliver_unloadWorld(theWorld, false)
+        if (!unloadResult.isSuccessful) {
+            Bukkit.broadcast("failed to delete due to: $unloadResult".mm(), "isop.isop")
+        } else {
+            println("starting delete")
+            Task.asyncDelayed((Ticks.TICKS_PER_SECOND * 10).toLong()) {
+                if (!theWorld.worldFolder.deleteRecursively())
+                    Bukkit.broadcast("failed to delete world folder".mm(), "isop.isop")
+                println("done deleting")
+            }
         }
     }
 }
